RAG_1/requirements.txt
# requirements.txt
PyPDF2
langchain
huggingfacehub
langchain-community
langchain-huggingface
langchain-core
gradio
python-dotenv
pandas